+++ src/assets/index.js
@@ -1,0 +1,2 @@
+export { default as imageExample } from './images/image_Example.jpg';
+export { default as image2 } from './images/image_Example2.jpg';
