--- src/config/route.js
@@ -1,37 +1,0 @@
-import { Home, Profile, Contact, Login, ProfileId } from '../pages';
-
-const routes = [
-  {
-    path: '/profile/:profileId',
-    component: ProfileId,
-    isPublic: false,
-  },
-  {
-    path: '/profile',
-    component: Profile,
-    isPublic: false,
-  },
-  {
-    path: '/contact',
-    component: Contact,
-    isPublic: true,
-  },
-  {
-    path: '/home',
-    component: Home,
-    isPublic: true,
-  },
-  {
-    path: '/login',
-    component: Login,
-    isPublic: true,
-  },
-
-  {
-    path: '/',
-    component: Home,
-    isPublic: true,
-  },
-];
-
-export default routes;
