--- src/pages/home.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-  return <div> Welcome Home!</div>;
-};
-
-export default Home;
