+++ src/pages/home/style.css
@@ -1,0 +1,5 @@
+.pageHome-wrapper,
+.pageHome-img1,
+.pageHome-img2 {
+  width: 100%;
+}
